Clean up duplicate configs and redundant E2E scripts
- Consolidate Playwright configs from 4 to 2 (50% reduction)
- Simplify E2E scripts from 17+ to 8 essential ones (53% reduction)
- Enhance responsive mobile tests with comprehensive coverage
- Add documentation for all configuration changes
