+++ docs/playwright-config-consolidation.md
@@ -1,0 +1,72 @@
+# Playwright Configuration Consolidation
+
+## Summary
+Reduced from 4 Playwright configurations to 2 focused configurations, eliminating redundancy and confusion.
+
+## Before (4 configs):
+- ❌ `playwright.config.ts` - Main config (basic)
+- ✅ `playwright.config.optimized.ts` - Enhanced config (actively used)
+- ✅ `playwright.config.dev.ts` - Ultra-fast development config
+- ❌ `playwright.config.parallel-future.ts` - Experimental (marked "DO NOT USE")
+
+## After (2 configs):
+- ✅ `playwright.config.ts` - Main config (promoted from optimized)
+- ✅ `playwright.config.dev.ts` - Ultra-fast development config
+
+## Changes Made
+
+### 1. Promoted Optimized Config
+- **Action:** Renamed `playwright.config.optimized.ts` → `playwright.config.ts`
+- **Reason:** The optimized config was already used by all specialized npm scripts
+- **Benefits:** Enhanced performance detection, better logging, optimized browser args
+
+### 2. Removed Redundant Configs
+- **Removed:** Old `playwright.config.ts` (basic version)
+- **Removed:** `playwright.config.parallel-future.ts` (experimental)
+- **Reason:** Superseded by optimized version / not ready for use
+
+### 3. Updated Package.json Scripts
+- **Changed:** Removed explicit `--config=playwright.config.optimized.ts` references
+- **Added:** New `test:e2e:dev-ultra` script for ultra-fast development
+- **Result:** All scripts now use the main config by default
+
+## Usage
+
+### Main Configuration (Default)
+```bash
+npm run test:e2e              # Standard E2E tests
+npm run test:e2e:fast         # Fast mode
+npm run test:e2e:fail-fast    # Fail fast mode
+npm run test:e2e:optimized    # Optimized mode
+npm run test:e2e:perf         # Performance logging
+```
+
+### Development Configuration (Ultra-Fast)
+```bash
+npm run test:e2e:dev-ultra    # Ultra-fast development (1s timeouts)
+```
+
+## Benefits
+
+1. **Reduced Complexity:** 50% fewer config files
+2. **Eliminated Confusion:** No more wondering which config to use
+3. **Better Defaults:** Main config now uses the enhanced optimized version
+4. **Maintained Flexibility:** Still have fast development option
+5. **Cleaner Package.json:** Removed redundant config specifications
+
+## Configuration Features
+
+### Main Config (`playwright.config.ts`)
+- Environment detection (CI, fast mode, optimized mode)
+- Enhanced performance logging
+- Worker optimization based on environment
+- Sequential execution for stability
+- Comprehensive project configurations
+- Advanced browser launch optimizations
+
+### Dev Config (`playwright.config.dev.ts`)
+- Ultra-aggressive timeouts (1s assertions, 10s tests)
+- Maximum parallelization
+- No retries (immediate feedback)
+- Minimal reporting for speed
+- Perfect for rapid development iteration
