+++ package.json
@@ -17,25 +17,14 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage --coverageReporters=lcov",
-    "test:e2e": "AUTH_ONLY=1 playwright test",
-    "test:e2e:fast": "PLAYWRIGHT_FAST=true playwright test --config=playwright.config.optimized.ts",
-    "test:e2e:fail-fast": "PLAYWRIGHT_FAIL_FAST=true playwright test --config=playwright.config.optimized.ts",
-    "test:e2e:dev": "PLAYWRIGHT_FAIL_FAST=true PLAYWRIGHT_FAST=true playwright test --config=playwright.config.optimized.ts",
-    "test:e2e:optimized": "PLAYWRIGHT_OPTIMIZED=true playwright test --config=playwright.config.optimized.ts",
-    "test:e2e:perf": "PLAYWRIGHT_PERF_LOG=true PLAYWRIGHT_OPTIMIZED=true playwright test --config=playwright.config.optimized.ts",
-    "test:e2e:parallel": "PLAYWRIGHT_FAST=true PLAYWRIGHT_SHARED_DATA=true playwright test --config=playwright.config.optimized.ts",
-    "test:e2e:profile": "PLAYWRIGHT_FAST=true playwright test --config=playwright.config.optimized.ts --reporter=line",
-    "test:e2e:validate": "node scripts/validate-config.js",
-    "test:e2e:all": "npm run test:e2e:gallery && npm run test:e2e:authenticated",
+    "test:e2e": "playwright test",
+    "test:e2e:dev": "playwright test --config=playwright.config.dev.ts",
+    "test:e2e:fast": "PLAYWRIGHT_FAST=true PLAYWRIGHT_FAIL_FAST=true playwright test",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:debug": "playwright test --debug",
-    "e2e:cleanup": "node scripts/e2e-cleanup.js",
-    "e2e:cleanup:all": "node scripts/e2e-cleanup.js --delete-user",
-    "e2e:cleanup:screenshots": "./scripts/cleanup-screenshots.sh",
     "test:e2e:report": "playwright show-report",
-    "test:e2e:auth": "playwright test e2e-tests/auth.spec.ts --project=chromium",
-    "test:e2e:gallery": "SKIP_AUTH=1 playwright test e2e-tests/gallery.spec.ts --project=chromium",
-    "test:e2e:authenticated": "AUTH_ONLY=1 playwright test e2e-tests/authenticated.spec.ts --project=authenticated"
+    "test:e2e:perf": "PLAYWRIGHT_PERF_LOG=true PLAYWRIGHT_OPTIMIZED=true playwright test",
+    "e2e:cleanup": "node scripts/e2e-cleanup.js && ./scripts/cleanup-screenshots.sh"
   },
   "dependencies": {
     "@auth/prisma-adapter": "^2.9.0",
